much more modular way to call a robot to perform a move

+++ csoko_thinker/include/csoko_thinker/csoko_solver.h
@@ -95,7 +95,7 @@ void performOneMove(Table &t, T_pos &r, T_pos &b, const char &c);
 /**
  *
  */
-void getMovementDelta(char c, size_t &x, size_t &y);
+void getMovementDelta(char c, int &x, int &y);
 
 
 /**
